wip: consolidate create route def
