+++ src/routing/route/resolve-route-options.ts
@@ -1,13 +1,7 @@
+import { extractBodyOptions } from '@/routing/route/create-route-definition';
 import type { RouteOptions } from '@/routing/route/route-options.type';
 import type { Request } from 'koa';
 import type { RouteDefinition } from './route-definition.type';
-
-export function resolveRouteOptions(options: RouteOptions): Pick<RouteDefinition, 'parseBody' | 'bodyOptions'> {
-  return {
-    parseBody: options.parseBody ?? true,
-    bodyOptions: extractBodyOptions(options),
-  };
-}
 
 export function mergeRouteOptions<TRequest extends Request>(
   route: RouteDefinition<TRequest>,
@@ -20,10 +14,4 @@ export function mergeRouteOptions<TRequest extends Request>(
       ...extractBodyOptions(options),
     },
   };
-}
-
-function extractBodyOptions(options: RouteOptions): RouteDefinition['bodyOptions'] {
-  const { parseBody: _parseBody, ...bodyOptions } = options;
-
-  return bodyOptions as RouteDefinition['bodyOptions'];
 }
